requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.9.4
future==0.17.1
idna==2.8
line-bot-sdk==1.12.1
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
